Made it not move "payload"

+++ install.lua
@@ -7,7 +7,7 @@ end
 fs.mkdir(sandDir)
 
 for _, v in pairs(fs.list("/")) do
-    if (v ~= sandDir) and (not fs.isReadOnly("/" .. v)) then
+    if (v ~= sandDir) and (v ~= payload) and (not fs.isReadOnly("/" .. v)) then
         fs.move("/" .. v, sandDir .. "/" .. v)
     end
     fs.move(sandDir, "/sand")
